Updated Design Document to better fit how the program has been written i.e added new instance variables added methods and changed how some methods were planned to be executed
Updated the EquationRunner.java class to get welcome the user and get their input as a string before extracting the x and y coordinates, the correct results are printed then the program prompts the user for a third point before returning the correct coordinates for the inputted x value.

Updated the LinearEquation class to format the slope as a fraction correctly as well as print the correct third point. The equation method was fixed to display the correct equation based on the conditions.